% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_it_2013.R
\docType{data}
\name{lfs_it_2013}
\alias{lfs_it_2013}
\title{Labour Force Survey Italy 2013 (sample)}
\format{
A tibble with 1,000 rows and 6 variables:
\describe{
\item{REFYEAR}{Reference year}
\item{SEX}{Sex}
\item{AGE}{Age class}
\item{STAPRO}{Professional status}
\item{HWACTUAL}{Hours actually worked}
\item{COEFF}{Weighting factor}
}
}
\source{
Eurostat Labour Force Survey
}
\usage{
lfs_it_2013
}
\description{
A sample of 1,000 observations from the Italian Labour Force Survey 2013
(Eurostat), provided for illustrative purposes.
}
\keyword{datasets}
