% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfs_it_2013_labels.R
\docType{data}
\name{lfs_it_2013_labels}
\alias{lfs_it_2013_labels}
\title{Labels for Labour Force Survey Italy 2013}
\format{
A tibble with the following columns:
\describe{
\item{var}{Variable name}
\item{value}{Code value}
\item{label}{Label description}
\item{lang}{Language}
\item{type}{Variable type}
\item{weight}{Weight variable flag}
\item{missing}{Missing value flag}
}
}
\usage{
lfs_it_2013_labels
}
\description{
Variable metadata and labels associated with
\code{lfs_it_2013}.
}
\keyword{datasets}
