% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_final_fit.R
\docType{data}
\name{rf_final_fit}
\alias{rf_final_fit}
\title{Internal Random Forest Model}
\format{
An object of class \code{workflow} containing a fitted random forest model.
}
\description{
A trained tidymodels workflow used internally for variable type prediction.
}
\keyword{internal}
