% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typerclass-package.R
\docType{package}
\name{typerclass-package}
\alias{typerclass}
\alias{typerclass-package}
\title{typerclass: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dassi-archive/typerclass}
  \item Report bugs at \url{https://github.com/dassi-archive/typerclass/issues}
}

}
\author{
\strong{Maintainer}: Domingo Scisci \email{domingo.scisci@unimib.it} (\href{https://orcid.org/0000-0003-1206-5007}{ORCID})

Authors:
\itemize{
  \item Giovanna De Santis \email{giovanna.desantis@unimib.it} (\href{https://orcid.org/0009-0002-2595-5906}{ORCID})
}

}
\keyword{internal}
